Auth: Create sub package and type alias to avoid circular dependency
