--- pkg/models/user_token.go
@@ -1,9 +1,0 @@
-package models
-
-type TokenRevokedError struct {
-	UserID                int64
-	TokenID               int64
-	MaxConcurrentSessions int64
-}
-
-func (e *TokenRevokedError) Error() string { return "user token revoked" }
