+++ pkg/services/auth/auth.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"net"
 
+	"github.com/grafana/grafana/pkg/models/usertoken"
 	"github.com/grafana/grafana/pkg/registry"
 	"github.com/grafana/grafana/pkg/services/quota"
 	"github.com/grafana/grafana/pkg/services/user"
@@ -23,6 +24,8 @@ type ActiveTokenService interface {
 var (
 	ErrUserTokenNotFound = errors.New("user token not found")
 )
+
+type TokenRevokedError = usertoken.TokenRevokedError
 
 // CreateTokenErr represents a token creation error; used in Enterprise
 type CreateTokenErr struct {
@@ -45,22 +48,7 @@ type TokenExpiredError struct {
 
 func (e *TokenExpiredError) Error() string { return "user token expired" }
 
-// UserToken represents a user token
-type UserToken struct {
-	Id            int64
-	UserId        int64
-	AuthToken     string
-	PrevAuthToken string
-	UserAgent     string
-	ClientIp      string
-	AuthTokenSeen bool
-	SeenAt        int64
-	RotatedAt     int64
-	CreatedAt     int64
-	UpdatedAt     int64
-	RevokedAt     int64
-	UnhashedToken string
-}
+type UserToken = usertoken.UserToken
 
 type RevokeAuthTokenCmd struct {
 	AuthTokenId int64 `json:"authTokenId"`
